+++ src/torchmetrics/functional/__init__.py
@@ -27,10 +27,10 @@ from torchmetrics.functional.classification.hamming import hamming_distance
 from torchmetrics.functional.classification.hinge import hinge_loss
 from torchmetrics.functional.classification.jaccard import jaccard_index
 from torchmetrics.functional.classification.matthews_corrcoef import matthews_corrcoef
-from torchmetrics.functional.classification.precision_at_fixed_recall import precision_at_fixed_recall
+from torchmetrics.functional.classification.precision_fixed_recall import precision_at_fixed_recall
 from torchmetrics.functional.classification.precision_recall import precision, recall
 from torchmetrics.functional.classification.precision_recall_curve import precision_recall_curve
-from torchmetrics.functional.classification.recall_at_fixed_precision import recall_at_fixed_precision
+from torchmetrics.functional.classification.recall_fixed_precision import recall_at_fixed_precision
 from torchmetrics.functional.classification.roc import roc
 from torchmetrics.functional.classification.specificity import specificity
 from torchmetrics.functional.classification.stat_scores import stat_scores
