+++ src/torchmetrics/functional/classification/recall_fixed_precision.py
